requirements.txt
gymnasium==1.0.0
stable-baselines3==2.5.0
tensorboard==2.19.0
torchjd==0.5.0
git+https://github.com/Farama-Foundation/Metaworld.git@master#egg=metaworld
